+++ FE/src/features/game/components/AnswerModal.tsx
@@ -1,5 +1,5 @@
 import Lottie from 'lottie-react';
-import AnswerBg from '../assets/lottie/answer_background.json';
+import AnswerBg from '@/assets/lottie/answer_background.json';
 import { useEffect, useState } from 'react';
 
 type AnswerModalProps = {
