+++ FE/src/features/game/components/Modal.tsx
