+++ FE/src/features/game/data/store/usePlayerStore.ts
@@ -1,6 +1,4 @@
-import { socketService } from '@/api/socket';
 import { create } from 'zustand';
-import { useRoomStore } from './useRoomStore';
 
 type Player = {
   playerId: string; // socketId
@@ -93,69 +91,3 @@ export const usePlayerStore = create<PlayerStore>((set) => ({
   },
   reset: () => set({ players: [], isHost: false, currentPlayerId: '', currentPlayerName: '' })
 }));
-
-socketService.on('joinRoom', (data) => {
-  const { addPlayers, setCurrentPlayerId } = usePlayerStore.getState();
-  const newPlayers = data.players.map((player) => ({
-    ...player,
-    playerScore: 0,
-    isAlive: true,
-    isAnswer: true
-  }));
-  addPlayers(newPlayers);
-  const socketId = socketService.getSocketId();
-  if (newPlayers.length > 0 && newPlayers[0].playerId === socketId) {
-    setCurrentPlayerId(socketId);
-  }
-});
-
-socketService.on('updatePosition', (data) => {
-  usePlayerStore.getState().updatePlayerPosition(data.playerId, data.playerPosition);
-});
-
-socketService.on('endQuizTime', (data) => {
-  const { players, setPlayers } = usePlayerStore.getState();
-  const { gameMode } = useRoomStore.getState();
-
-  setPlayers(
-    data.players.map((p) => {
-      const _p = players.find((e) => e.playerId === p.playerId);
-      return {
-        playerId: String(p.playerId),
-        playerName: _p?.playerName || '',
-        playerPosition: _p?.playerPosition || [0, 0],
-        playerScore: p.score,
-        isAnswer: p.isAnswer,
-        isAlive: _p?.isAlive || false
-      };
-    })
-  );
-
-  // 서바이벌 모드일 경우 3초 뒤에 탈락한 플레이어를 보이지 않게 한다.
-  if (gameMode === 'SURVIVAL') {
-    setTimeout(() => {
-      const { players, setPlayers } = usePlayerStore.getState();
-
-      setPlayers(
-        players.map((p) => {
-          return {
-            ...p,
-            isAlive: p.isAlive && p?.isAnswer
-          };
-        })
-      );
-    }, 3000);
-  }
-});
-
-socketService.on('endGame', (data) => {
-  usePlayerStore.getState().setIsHost(data.hostId === socketService.getSocketId());
-});
-
-socketService.on('exitRoom', (data) => {
-  usePlayerStore.getState().removePlayer(data.playerId);
-});
-
-socketService.on('disconnect', () => {
-  usePlayerStore.getState().reset();
-});
