+++ FE/src/features/game/index.ts
@@ -1,0 +1,5 @@
+import './data/socketListener.ts';
+
+export * from './pages/GamePage.tsx';
+export * from './pages/GameSetupPage.tsx';
+export * from './pages/PinPage.tsx';
