+++ FE/src/features/game/utils/nickname.ts
