+++ FE/src/features/lobby/index.ts
@@ -1,0 +1,1 @@
+export * from './GameLobbyPage';
